`GaussianBaseParser`: Recursively check for `nan` and `inf` (#32)
* `GaussianBaseParser`: Recursively check for `nan` and `inf`

The data parsed by the `cclib` library may contain numpy arrays with
`nan` and `inf` values. These cannot be stored in a `Data` node, as they
are not JSON-serializable.

The parser already contained an ad-hoc solution for these values that
could occur in the `scfvalues` key, however, they can occur also in
other places in the dictionary. The solution is generalized and the
entire parsed dictionary is recursively scanned for numpy arrays
containing `nan` and `inf` values. If found, the `num_to_nan` function
of numpy is used to convert them. The default replace values are used
meaning `nan` is converted to `0.0` and `-inf/inf` is converted to a
very large negative/positive number. The arrays are changed in place to
reduce the memory footprint.

A test is added for the parser. The reference output file contains values
that when parsed, the `scfvalues` dictionary contains at least one `nan`
value. The reference output file comes from an actual calculation but
was heavily trimmed down. The original file was multiple MBs.

The `aiida_gaussian/tests` folder is added to the exclude list of the
`yamlfmt` pre-commit hook as that formats the reference files
differently then the `pytest-regressions` plugin causing the tests to
fail.

* Address PR review comments