Tests: Automate calculation test reference generation and checking
The reference file for the single `GaussianCalculation` test was
generated manually and read manually to compare to the one created by
running the calcjob plugin. This is replaced by using the fixture
`file_regression` which not only automates this, if the reference file
needs to change due to a change in the code, it can be automatically
generated by running the tests with the `--force-regen` flag.