`GaussianCalculation`: Do no limit kwargs for `GaussianInput`
The `GaussianCalculation` is rendering the input file from the
`parameters` input dictionary through the `GaussianInput` utility from
`pymatgen`. However, it was hard-coding which keywords could be
specified through the parameters. For example, the `gen_basis` keyword
of `GaussianInput` was ignored. Here we simply expand the entire
parameters dictionary as it is provided through the inputs, after
setting some relevant defaults.