`GaussianCalculation`: Ensure `basis_set` parameter is respected
The `GaussianInput` class automatically sets `basis_set` to `Gen` if
`gen_basis` is specified. However, if the user explicitly set `basis_set`
to ``genecp``, we should respect that, since if the custom basis set
definition contains effective core potentials, running with `genecp` is
significantly faster than just running with `gen`, which seems to ignore
the ECPs.