`GaussianCalculation`: Do no limit kwargs for `GaussianInput` (#34)
* Tests: Automate calculation test reference generation and checking

The reference file for the single `GaussianCalculation` test was
generated manually and read manually to compare to the one created by
running the calcjob plugin. This is replaced by using the fixture
`file_regression` which not only automates this, if the reference file
needs to change due to a change in the code, it can be automatically
generated by running the tests with the `--force-regen` flag.

* `GaussianCalculation`: Do no limit kwargs for `GaussianInput`

The `GaussianCalculation` is rendering the input file from the
`parameters` input dictionary through the `GaussianInput` utility from
`pymatgen`. However, it was hard-coding which keywords could be
specified through the parameters. For example, the `gen_basis` keyword
of `GaussianInput` was ignored. Here we simply expand the entire
parameters dictionary as it is provided through the inputs, after
setting some relevant defaults.